% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_down.R
\name{step_down}
\alias{step_down}
\title{Step-Down algorithm to control the FDP}
\usage{
step_down(target_scores, decoy_scores, alpha, gamma, randomised = T)
}
\arguments{
\item{target_scores}{a vector of numeric values representing
the target scores}

\item{decoy_scores}{a vector of numeric values representing the corresponding
decoy scores}

\item{alpha}{the FDP control threshold, must be a numeric value between 0 and 1}

\item{gamma}{the confidence level will be 1 - gamma}

\item{randomised}{boolean}
}
\value{
It returns a threshold and a list of indices of the dicvoeries
}
\description{
Step-Down algorithm to control the FDP
}
\examples{
x <- rnorm(1000, 1, 1)
y <- rnorm(1000)
target_scores <- apply(cbind(x, y), 1, max)
decoy_scores <- rnorm(1000)
step_down(target_scores, decoy_scores, 0.2, 0.1)
}
